# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A FizzBuzz REST API that returns the classic sequence for caller-supplied divisors and words, and reports the most requested combination.",
        "title": "FizzBuzz API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/fizzbuzz": {
            "get": {
                "description": "Returns a fizzbuzz list from 1 to limit: multiples of int1 become str1, multiples of int2 become str2, multiples of both become str1+str2.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fizzbuzz"
                ],
                "summary": "FizzBuzz",
                "parameters": [
                    {
                        "maximum": 1000,
                        "minimum": 1,
                        "type": "integer",
                        "description": "First divisor",
                        "name": "int1",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 1000,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Second divisor",
                        "name": "int2",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 1000,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Upper bound, inclusive",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "fizz",
                        "description": "Replacement for int1 multiples",
                        "name": "str1",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "buzz",
                        "description": "Replacement for int2 multiples",
                        "name": "str2",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/fizzbuzz/stats": {
            "get": {
                "description": "Returns the most requested fizzbuzz parameters and how many times they have been asked for. Responds with {\"message\": \"no requests yet\"} when nothing has been recorded.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fizzbuzz"
                ],
                "summary": "FizzBuzz statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    }
}